refactor: colocate config unit tests
